pyproject.toml
[project]
name = "nonebot-plugin-dst-qq"
version = "0.4.0"
description = "基于 NoneBot2 的饥荒管理平台 (DMP) QQ 机器人插件，支持游戏信息查询、命令执行和消息互通功能"
readme = "README.md"
requires-python = ">=3.9, <4.0"
authors = [
    {name = "uitok", email = "ui_101@qq.com"}
]
license = {text = "MIT"}
keywords = ["nonebot", "nonebot2", "plugin", "dst", "dmp", "饥荒", "don't starve together", "qq-bot", "message-exchange"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Games/Entertainment",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
]
dependencies = [
    "nonebot2>=2.4.0",
    "nonebot-adapter-onebot>=2.4.0",
    "nonebot-plugin-alconna>=0.22.0",
    "nonebot-plugin-localstore>=0.6.0",
    "nonebot-plugin-htmlrender>=0.3.0",
    "httpx>=0.24.0",
    "pydantic",
    "aiosqlite>=0.19.0",
    "pillow>=10.0.0",
    "watchdog>=3.0.0",
    "playwright>=1.40.0"
]

[project.urls]
Homepage = "https://github.com/uitok/nonebot-plugin-dst-qq"
Repository = "https://github.com/uitok/nonebot-plugin-dst-qq"
Documentation = "https://github.com/uitok/nonebot-plugin-dst-qq#readme"
"Bug Tracker" = "https://github.com/uitok/nonebot-plugin-dst-qq/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nonebot_plugin_dst_qq"]
exclude = [
    "nonebot_plugin_dst_qq/app_config.json",
    "nonebot_plugin_dst_qq/app_config.backup.json",
    "nonebot_plugin_dst_qq/*.json.bak",
]

[tool.hatch.build.targets.sdist]
include = [
    "/nonebot_plugin_dst_qq",
    "/README.md",
    "/LICENSE",
]
exclude = [
    "/nonebot_plugin_dst_qq/app_config.json",
    "/nonebot_plugin_dst_qq/app_config.backup.json",
    "/nonebot_plugin_dst_qq/*.json.bak",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_alconna", "nonebot_plugin_localstore", "nonebot_plugin_apscheduler", "nonebot_plugin_htmlrender"]
plugin_dirs = ["nonebot_plugin_dst_qq"]
builtin_plugins = []

